requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
pydantic
pydantic-settings
python-dotenv
requests
slowapi
twilio==8.12.0 